LPS-55531 Move file system stores to its module
