+++ modules/portal/portal-store-filesystem/src/com/liferay/portal/store/filesystem/AdvancedFileSystemStore.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.store;
+package com.liferay.portal.store.filesystem;
 
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.util.CharPool;
