+++ modules/portal/portal-store-filesystem/test/integration/com/liferay/portal/store/filesystem/test/AdvancedFileSystemStoreTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.store;
+package com.liferay.portal.store.filesystem.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
@@ -23,19 +23,18 @@ import org.junit.ClassRule;
 import org.junit.Rule;
 
 /**
- * @author Preston Crary
+ * @author Vilmos Papp
  */
-public class FileSystemStoreTest extends BaseStoreTestCase {
+public class AdvancedFileSystemStoreTest extends BaseStoreTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE);
+		new AggregateTestRule(new LiferayIntegrationTestRule());
 
 	@Override
 	protected Store getStore() {
-		return new FileSystemStore();
+		return new AdvancedFileSystemStore();
 	}
 
 }
