+++ portal-impl/src/portal.properties
@@ -8985,14 +8985,6 @@
     #dl.store.antivirus.impl=com.liferay.portlet.documentlibrary.antivirus.ClamAntivirusScannerImpl
 
     #
-    # Set the name of a class that implements
-    # com.liferay.portlet.documentlibrary.store.Store. The
-    # document library server will use this to persist documents.
-    #
-    #dl.store.impl=com.liferay.portlet.documentlibrary.store.AdvancedFileSystemStore
-    dl.store.impl=com.liferay.portlet.documentlibrary.store.FileSystemStore
-
-    #
     # AdvancedFileSystemStore
     #
     dl.store.advanced.file.system.root.dir=${liferay.home}/data/document_library
